requirements.txt
gradio
python-dotenv
langgraph
typing-extensions
langchain
langchain-google-genai
Pillow